# Python Sample/fatsever/API.yaml
openapi: 3.0.3
info:
  title: FastAPI Project
  version: "1.0.0"
  description: >
    A FastAPI application that supports file uploads, video downloads, and basic HTML interface.
servers:
  - url: http://localhost:8000
paths:
...
  /upload/file:
    post:
      summary: Upload a file
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                  status:
                    type: string
                required:
                  - filename
                  - status
        "400":
          description: Bad Request
...